query_based_report/requirements.txt
# LangChain & friends (stable inter-compatible set)
langchain==0.2.14
langchain-core==0.2.32
langchain-community==0.2.12
langchain-experimental==0.0.64


# Gemini
langchain-google-genai==1.0.6
google-generativeai==0.5.4

# Vector DB + embeddings
chromadb==0.5.4
sentence-transformers==2.2.2
torch>=2.0.0

# DB & server
SQLAlchemy==2.0.36
pymysql==1.1.1
mysqlclient==2.2.4
python-dotenv==1.0.1
streamlit==1.39.0
pandas==2.2.3
tabulate==0.9.0
tqdm==4.66.5
